fix: simplified cart
